+++ app/src/main/java/com/kuit/kuit6android/navigation/Route.kt
@@ -16,12 +16,14 @@ sealed interface Route {
         val searchKeyword: String,
     ) : Route
 
+    // Navigation Graph를 그룹으로 묶는 역할
     @Serializable
     data object SearchNestedGraphRoute : Route
 
     @Serializable
     data object Favorite : Route
 
+    // 데이터를 담고 이동해야 하는 라우트
     @Serializable
     data class DetailPage(
         val imageId : Int,
@@ -31,6 +33,7 @@ sealed interface Route {
         val reviewCount : String
     ) : Route
 
+
     @Serializable
     data object FavoriteNestedGraphRoute : Route
 
@@ -39,6 +42,9 @@ sealed interface Route {
 
     @Serializable
     data object MyEats : Route
+
+    @Serializable
+    data object MyEatsFavorite : Route
 
     @Serializable
     data object MyEatsNestedGraphRoute : Route
